Fixing Problems
The first 2 debug statements are completely irrelevant, remove them and move on. Debug 3 is changed to debug 1, and will now show the value held in board[].  Sadly, the simple action of pushing the Change Opponent button can cause the board to fill up instantaneously with garbage moves. Look into fixing possible recursion problems (i.e. deep copies vs. shallow copies). Consider adding a new input variable for evaluateMove() that signifies what move makes the child differ from the parent (This could lead to a way to get around the deep/shallow variable problem). score[] is fine the way it is; null wont cause any problems, and the order needs to be maintained properly, so it's necessary.