CPU Fix
Hopefully this will solve the problem.  Apparently every variable created with "var ___ = ___;" (including the iteration variable used in for-loops) are actually global variables, and only function arguments act as local variables.  Because of this, several variables used in evaluateMove() (including major variables like board, i, depth, and scores) would instantaneously jumble when evaluateMove() was called.  To fix this issue, I added an array stack[], that would properly retain data used in recursive calls, most importantly scores & i.  For board, I decided to ignore fiddling with arguments and just manually reverted excess changes.  depth is still just an argument, so it will work properly if I remove the use of newDepth.  
In addition to these major changes, I cleaned up some of the unnecessary arguments I had been previously using, like the move & brd variables in evaluateMove(), and the brd argument in checkVictory().  The code is now a bit more specialized, but it's OK because all of this code will only be used on this one page.  
Next Steps: Tidy up the comments, and remove the debug portions.  (if this works)